refactor(schema-api): extract DictionaryApi trait
Extract DictionaryApi trait from monolithic SchemaApi to improve code
organization and maintainability. This continues the ongoing schema API
refactoring effort by moving dictionary-related operations into their
own focused trait interface.
