+++ src/meta/api/src/lib.rs
@@ -23,6 +23,7 @@ mod data_mask_api;
 mod data_mask_api_impl;
 mod database_api;
 pub mod database_util;
+pub mod dictionary_api;
 pub mod index_api;
 pub mod kv_app_error;
 pub mod kv_pb_api;
@@ -49,6 +50,7 @@ pub(crate) mod sequence_nextval_impl;
 pub use catalog_api::CatalogApi;
 pub use data_mask_api::DatamaskApi;
 pub use database_api::DatabaseApi;
+pub use dictionary_api::DictionaryApi;
 pub use index_api::IndexApi;
 pub use row_access_policy_api::RowAccessPolicyApi;
 pub use schema_api::SchemaApi;
